requirements-docker.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.35
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
# Optional (depending on your orchestrator):
langchain==0.2.11
langchain-openai==0.1.17
langchain-community==0.2.10
tiktoken==0.7.0
